# Performance Improvement in tds and ucs2 parsing